requirements.txt
Django>=4.2,<4.3
python-dotenv
requests==2.25.1
mysqlclient
gunicorn
djangorestframework
django-cors-headers